+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - [compose] Bumped Elasticsearch to 7.17.26.
 - [crawler] Gitlab - make Notes author attribute as nullable
 - [crawler] Gitlab - lower the bulk amount of MR objects to fetch from 100 to 25
+- [crawler] Gitlab - fetch the approval status
 
 ### Removed
 ### Fixed